refactor: Move TransactionTableHead to TransactionHeader
