+++ src/ducks/transactions/TransactionHeader.jsx
@@ -8,11 +8,42 @@ import { Breadcrumb } from 'components/Breadcrumb'
 import { ConnectedSelectDates } from 'components/SelectDates'
 import { AccountSwitch } from 'ducks/account'
 import TransactionSelectDates from 'ducks/transactions/TransactionSelectDates'
-import { TransactionTableHead } from './Transactions'
 import flag from 'cozy-flags'
 import HistoryChart from 'ducks/balance/HistoryChart'
+import { Table } from 'components/Table'
 
 import styles from './TransactionsPage.styl'
+import transactionsStyles from './Transactions.styl'
+
+const sDate = transactionsStyles['bnk-op-date']
+const sDesc = transactionsStyles['bnk-op-desc']
+const sAmount = transactionsStyles['bnk-op-amount']
+const sAction = transactionsStyles['bnk-op-action']
+
+const TableHeadDesktop = ({ t, mainColumnTitle }) => (
+  <thead>
+    <tr>
+      <td className={sDesc}>{mainColumnTitle}</td>
+      <td className={sDate}>{t('Transactions.header.date')}</td>
+      <td className={sAmount}>{t('Transactions.header.amount')}</td>
+      <td className={sAction}>{t('Transactions.header.action')}</td>
+    </tr>
+  </thead>
+)
+
+const transactionTableHeadStyle = {
+  background: 'white',
+  marginTop: '1rem',
+  marginLeft: '-2rem',
+  marginRight: '-2rem'
+}
+export const TransactionTableHead = translate()(props => (
+  <div style={transactionTableHeadStyle}>
+    <Table className={styles['TransactionTable']}>
+      <TableHeadDesktop t={props.t} mainColumnTitle={props.mainColumnTitle} />
+    </Table>
+  </div>
+))
 
 const historyChartMargin = {
   top: 10,
