+++ src/ducks/transactions/Transactions.jsx
@@ -50,17 +50,6 @@ const sDesc = styles['bnk-op-desc']
 const sAmount = styles['bnk-op-amount']
 const sAction = styles['bnk-op-action']
 
-const TableHeadDesktop = ({ t, mainColumnTitle }) => (
-  <thead>
-    <tr>
-      <td className={sDesc}>{mainColumnTitle}</td>
-      <td className={sDate}>{t('Transactions.header.date')}</td>
-      <td className={sAmount}>{t('Transactions.header.amount')}</td>
-      <td className={sAction}>{t('Transactions.header.action')}</td>
-    </tr>
-  </thead>
-)
-
 class _TableTrDesktop extends PureComponent {
   onSelectTransaction = () =>
     this.props.selectTransaction(this.props.transaction)
@@ -214,20 +203,6 @@ _TableTrNoDesktop.propTypes = {
 }
 
 export const TableTrNoDesktop = translate()(_TableTrNoDesktop)
-
-const transactionTableHeadStyle = {
-  background: 'white',
-  marginTop: '1rem',
-  marginLeft: '-2rem',
-  marginRight: '-2rem'
-}
-export const TransactionTableHead = translate()(props => (
-  <div style={transactionTableHeadStyle}>
-    <Table className={styles['TransactionTable']}>
-      <TableHeadDesktop t={props.t} mainColumnTitle={props.mainColumnTitle} />
-    </Table>
-  </div>
-))
 
 const groupByDateAndSort = transactions => {
   const byDate = groupBy(transactions, x => format(x.date, 'YYYY-MM-DD'))
